pyproject.toml
[project]
name = "pgstream"
version = "0.1.0"
description = "Production-grade Python SDK for Postgres CDC: watch tables via logical replication and sync changes to vector stores (pgvector, Qdrant)."
readme = "README.md"
authors = [
    { name = "Kamalesh-Kavin", email = "kamalesh.s7316@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
pgvector = [
    "asyncpg>=0.30.0",
]
qdrant = [
    "qdrant-client>=1.14.0",
]
all = [
    "asyncpg>=0.30.0",
    "qdrant-client>=1.14.0",
]

[build-system]
requires = ["uv_build>=0.10.10,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "asyncpg>=0.30.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require a live Postgres instance (deselect with -m 'not integration')",
]
